feat: Section component responsiveness
