+++ src/components/App/App.js
@@ -1,0 +1,13 @@
+import './App.css';
+import Section from '../Section/Section'
+
+function App() {
+  return (
+    <>
+      <Section type='products' layout='two-columns' title='Wide choice of transport robots'></Section>
+      <Section type='benefits' layout='four-columns' title='Benefits'></Section>
+    </>
+  );
+}
+
+export default App;
