+++ src/components/Section/Section.css
@@ -1,0 +1,91 @@
+.section {
+  background-color: #f9f9f9;
+  border-radius: 14px;
+  display: flex;
+  padding: 110px 100px 110px 100px;
+  margin: 10px;
+  box-sizing: content-box;
+}
+
+.section_layout_two-columns {
+  flex-direction: row;
+  min-height: 554px;
+}
+
+.section_layout_four-columns {
+  flex-direction: column;
+  min-height: 500px;
+}
+
+.section__title {
+  /* font-family: 'Inter', Arial, Helvetica, sans-serif; */
+  font-family: Arial;
+  font-size: 34px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 43px;
+  letter-spacing: 0em;
+  text-align: left;
+  margin: 0;
+  padding: 0;
+}
+
+.placeholder_two-columns {
+  width: 575px;
+  height: 324px;
+  background-color: brown;
+
+  margin: 0;
+  padding: 0;
+}
+
+.placeholder_four-columns {
+  width: 100%;
+  height: 151px;
+  margin: 0;
+  padding: 0;
+  display: flex;
+  justify-content: space-between;
+}
+
+.placeholder__card {
+  width: 235px;
+  height: 151px;
+  background-color: brown;
+  margin-bottom: 50px;
+}
+
+@media screen and (max-width: 768px) {
+  .section {
+    padding: 110px 0 0 0;
+    /* will need to set other margins/padding on modifiers */
+  }
+
+  .section_layout_two-columns {
+    flex-direction: column;
+  }
+
+  .section_layout_four-columns {
+    min-height: 767px;
+  }
+
+  .placeholder_two-columns {
+    width: auto;
+  }
+
+  .placeholder_four-columns {
+    flex-wrap: wrap;
+  }
+}
+
+@media screen and (max-width: 500px) {
+  .placeholder__card {
+    width: calc((100vw - 30px) / 2);
+  }
+}
+
+@media screen and (max-width: 425px) {
+  .placeholder__card {
+    width: calc(100vw - 30px);
+  }
+}
